refactor(app): extract modal/palette/dialog host JSX into ModalHostLayer
- Move the ~40 ErrorBoundary -> Suspense -> Lazy palette/dialog/
  confirm/onboarding host blocks out of AppInner's JSX return into a
  single ModalHostLayer component, one boundary + Suspense per lazy
  child preserved exactly (fault isolation unchanged)
- AppInner now composes hooks, then <AppLayout> + <ModalHostLayer>
- src/App.tsx: 1795 -> 636 lines
- No logic changes
